BLGMIGRATION-866 clean: move package
