+++ core/src/main/scala/com/nulabinc/backlog/migration/common/deserializers/Deserializer.scala
@@ -1,0 +1,5 @@
+package com.nulabinc.backlog.migration.common.deserializers
+
+trait Deserializer[A, B] {
+  def deserialize(a: A): B
+}
