--- core/src/main/scala/com/nulabinc/backlog/migration/common/domain/mappings/Deserializer.scala
@@ -1,5 +1,0 @@
-package com.nulabinc.backlog.migration.common.domain.mappings
-
-trait Deserializer[A, B] {
-  def deserialize(a: A): B
-}
