--- core/src/main/scala/com/nulabinc/backlog/migration/common/domain/mappings/Serializer.scala
@@ -1,5 +1,0 @@
-package com.nulabinc.backlog.migration.common.domain.mappings
-
-trait Serializer[A, B] {
-  def serialize(a: A): B
-}
