+++ core/src/main/scala/com/nulabinc/backlog/migration/common/formatters/Formatter.scala
@@ -1,4 +1,4 @@
-package com.nulabinc.backlog.migration.common.domain.mappings
+package com.nulabinc.backlog.migration.common.formatters
 
 trait Formatter[A] {
   def format(value: A): (String, String)
