+++ core/src/main/scala/com/nulabinc/backlog/migration/common/serializers/Serializer.scala
@@ -1,0 +1,5 @@
+package com.nulabinc.backlog.migration.common.serializers
+
+trait Serializer[A, B] {
+  def serialize(a: A): B
+}
