+++ core/src/main/scala/com/nulabinc/backlog/migration/common/service/PriorityMappingFileService.scala
@@ -4,8 +4,11 @@ import java.nio.file.Path
 
 import cats.Monad
 import cats.implicits._
+import com.nulabinc.backlog.migration.common.deserializers.Deserializer
 import com.nulabinc.backlog.migration.common.domain.mappings._
 import com.nulabinc.backlog.migration.common.dsl.{ConsoleDSL, StorageDSL}
+import com.nulabinc.backlog.migration.common.formatters.Formatter
+import com.nulabinc.backlog.migration.common.serializers.Serializer
 import com.nulabinc.backlog4j.Priority
 import org.apache.commons.csv.CSVRecord
