+++ core/src/main/scala/com/nulabinc/backlog/migration/common/service/StatusMappingFileService.scala
@@ -4,9 +4,12 @@ import java.nio.file.Path
 
 import cats.Monad
 import cats.implicits._
+import com.nulabinc.backlog.migration.common.deserializers.Deserializer
 import com.nulabinc.backlog.migration.common.domain.BacklogStatuses
 import com.nulabinc.backlog.migration.common.domain.mappings._
 import com.nulabinc.backlog.migration.common.dsl.{ConsoleDSL, StorageDSL}
+import com.nulabinc.backlog.migration.common.formatters.Formatter
+import com.nulabinc.backlog.migration.common.serializers.Serializer
 import org.apache.commons.csv.CSVRecord
 
 private case class MergedStatusMapping[A](mergeList: Seq[StatusMapping[A]], addedList: Seq[StatusMapping[A]])
