+++ core/src/main/scala/com/nulabinc/backlog/migration/common/service/UserMappingFileService.scala
@@ -5,9 +5,12 @@ import java.nio.file.Path
 import cats.Monad
 import cats.implicits._
 import com.nulabinc.backlog.migration.common.conf.BacklogApiConfiguration
+import com.nulabinc.backlog.migration.common.deserializers.Deserializer
 import com.nulabinc.backlog.migration.common.domain.BacklogUser
 import com.nulabinc.backlog.migration.common.domain.mappings._
 import com.nulabinc.backlog.migration.common.dsl.{ConsoleDSL, StorageDSL}
+import com.nulabinc.backlog.migration.common.formatters.Formatter
+import com.nulabinc.backlog.migration.common.serializers.Serializer
 import org.apache.commons.csv.CSVRecord
 
 private case class MergedUserMapping[A](mergeList: Seq[UserMapping[A]], addedList: Seq[UserMapping[A]])
